planning_interface: Set is_diff true for empty start state
Executing a motion without setting the start state of the robot like
this:
```
group_arm.setNamedTarget("start_grab_pose");
success = group_arm.move();
```
throws the error: Execution of motions should always start at the robot's
current state. Ignoring the state supplied as start state in the motion
planning request.

The problem is, when considered_start_state_ is null, every data field of the start_state
in the submitted MotionPlanRequest is 0 or false. But we need is_diff to be
true, because otherwise move_group will not consider its current state as
actual start state without complaining.