rviz display: disable Query Start State by default
This property is "only" used by people playing around with planning internals,
e.g. the people maintaining the code..

The "average" user does not expect to be able to manipulate the start state of the motion-request
te makes from rviz - it doesn't make sense if you want to actually move the robot.
Therefore, disable the start-state by default to reduce the initial complexity
the average user is confronted with.