rviz gui: disable execution if custom start state is selected
This prevents the execution of plans when the start state might not be the current state of the robot.

This also prevents the execution of some valid plans if the user explicitly set the (custom)
start state to match the current state beforehand, but I believe the benefits are worth it.