fixed unnecessary waiting
If the robot doesn't move, an update in CSM is not forwarded to the planning scene.
Hence, we always wait until timeout for a recent last_robot_motion_time_.
If we ensure that an update, if it is triggered by CSM, is directly
passed to the scene, we can early return true.

This partially reverts commit 907980a392d352f5e2ebb6a0b6906c85d9c7d72c.